// client/src/generated-ts-client.ts
export interface Todo {
    id?: string;
    title?: string;
    description?: string;
    priority?: number;
    isdone?: boolean;
}

export interface CreateTodoDto {
    priority?: number;
    title?: string;
    description?: string;
}